+++ crates/zng-view/src/lib.rs
@@ -11,8 +11,8 @@
 //!
 //! ```toml
 //! [dependencies]
-//! zng = "0.14.0"
-//! zng-view = "0.8.0"
+//! zng = "0.14.1"
+//! zng-view = "0.8.1"
 //! ```
 //!
 //! Then call `zng::env::init` before any other code in `main` to setup a view-process that uses